Feat: Implemented Equivalent Schedule Pruning

+++ src/main/java/scheduler/schedulers/parallel/ParallelSchedulerDynamic.java
@@ -58,14 +58,14 @@ public class ParallelSchedulerDynamic extends AStarScheduler {
             }
 
             for (NodeModel node : getAvailableNodes(state)) {
-                for (int processor = 0; processor < processors; processor++) {
+                for (byte processor = 0; processor < processors; processor++) {
                     expandState(state, node, processor);
                 }
             }
         }
     }
 
-    private void expandState(StateModel state, NodeModel node, int processor) {
+    private void expandState(StateModel state, NodeModel node, byte processor) {
         if (isFirstAvailableNode(state, node)) {
             return;
         }
@@ -188,13 +188,13 @@ public class ParallelSchedulerDynamic extends AStarScheduler {
                     continue;
                 }
 
-                for (int processor = 0; processor < processors; processor++) {
+                for (byte processor = 0; processor < processors; processor++) {
                     expandState(state, node, processor);
                 }
             }
         }
 
-        private void expandState(StateModel state, NodeModel node, int processor) {
+        private void expandState(StateModel state, NodeModel node, byte processor) {
             StateModel nextState = state.clone();
 
             int earliestStartTime = getEarliestStartTime(state, node, processor);
